Specify container item for MilkBucket replacement
Forgive me if this fails to compile, as I've done it (horribly) from the browser in a rush. 

This fixes IF milk buckets used in recipes having the bucket deleted when used in recipes rather than returned as per usual. Item.containerItem is actually defined through properties rather than through a function, and the default BucketItem does (obviously) no longer specify this.