Fix Resourceful Furnace not filling with Essense
Without this change the Resourceful Furnace is not filling up with Essense at all.

To be honest, I just looked at the Plant Gatherer and copied from there, as that one is actually working and filling with Sludge. 
I'm not sure if I broke something else by doing this, but I hope this at least helps you guys with a place to start.

I have no experience with mod development, but it was nice to see how easy it is to set up this project in Intellij. Great job on that!